Changed saving to save all routes
